requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.6.1
nsfw==0.3.2
Pillow==6.0.0
pyaes==1.6.1
pyAesCrypt==0.4.2
pycparser==2.19
Pyrogram==0.12.0
PySocks==1.6.8
six==1.12.0
TgCrypto==1.1.1
